Refactor authentication API integration and enhance trip creation components
- Replaced the existing API integration with a new backend API (`beApi`) across authentication-related components including login, signup, and password reset.
- Updated the `RootLayout` to include new state management for user readiness and health status.
- Removed deprecated `TripPlanner` component and its associated modal, streamlining the trip creation process.
- Adjusted imports in `SectionHeader` and `TripItemCard` to reflect the new component structure.
- Cleaned up unused API files and related data sources to improve codebase maintainability.
