+++ .vscode/c_cpp_properties.json
@@ -1,0 +1,14 @@
+{
+  "configurations": [
+    {
+      "name": "windows-gcc-x64",
+      "includePath": ["${workspaceFolder}/**"],
+      "compilerPath": "gcc",
+      "cStandard": "${default}",
+      "cppStandard": "${default}",
+      "intelliSenseMode": "windows-gcc-x64",
+      "compilerArgs": [""]
+    }
+  ],
+  "version": 4
+}
