--- components/CreateTripComponents/TripPlanner/AddItemModal.tsx
@@ -1,91 +1,0 @@
-import React from 'react'
-import { Modal, StyleSheet, Text, TouchableOpacity, View } from 'react-native'
-import { AddItemModalProps } from '../TripPlanner'
-
-export const AddItemModal: React.FC<AddItemModalProps> = ({
-  visible,
-  selectedTime,
-  onClose,
-  onConfirm,
-}) => {
-  return (
-    <Modal
-      transparent={false}
-      visible={visible}
-      onRequestClose={onClose}
-      animationType="slide"
-    >
-      <View style={styles.modalContainer}>
-        <View style={styles.modalContent}>
-          <Text style={styles.modalTitle}>Add New Item</Text>
-          <Text>Add item to {selectedTime?.toUpperCase()}</Text>
-
-          <View style={styles.modalButtons}>
-            <TouchableOpacity
-              style={[styles.modalButton, styles.cancelButton]}
-              onPress={onClose}
-            >
-              <Text style={styles.buttonText}>Cancel</Text>
-            </TouchableOpacity>
-
-            <TouchableOpacity
-              style={[styles.modalButton, styles.confirmButton]}
-              onPress={onConfirm}
-            >
-              <Text style={[styles.buttonText, styles.confirmButtonText]}>
-                Add
-              </Text>
-            </TouchableOpacity>
-          </View>
-        </View>
-      </View>
-    </Modal>
-  )
-}
-
-const styles = StyleSheet.create({
-  modalContainer: {
-    flex: 1,
-    width: '100%',
-    height: '100%',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'white',
-  },
-  modalContent: {
-    width: '100%',
-    height: '100%',
-    padding: 20,
-    alignItems: 'center',
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 15,
-  },
-  modalButtons: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    width: '100%',
-    marginTop: 20,
-  },
-  modalButton: {
-    padding: 10,
-    borderRadius: 8,
-    width: '45%',
-    alignItems: 'center',
-  },
-  cancelButton: {
-    backgroundColor: '#f3f3f3',
-  },
-  confirmButton: {
-    backgroundColor: '#563D30',
-  },
-  buttonText: {
-    fontSize: 16,
-    color: '#333',
-  },
-  confirmButtonText: {
-    color: 'white',
-  },
-})
