+++ components/CreateTripComponents/TripPlanner/SectionHeader.tsx
@@ -1,7 +1,7 @@
 import FontAwesome6 from '@expo/vector-icons/FontAwesome6'
 import React from 'react'
 import { StyleSheet, Text, TouchableOpacity, View } from 'react-native'
-import { SectionHeaderProps } from '../TripPlanner'
+import { SectionHeaderProps } from '../../../features/trips/presentation/components/TripPlanner'
 
 export const SectionHeader: React.FC<SectionHeaderProps> = ({
   time,
