+++ constants/utilConstants.ts
@@ -1,0 +1,2 @@
+export const EMPTY_STRING = ''
+export const INDENT_LEVEL = 2
