--- domain/repositories/PlaceRepository.ts
@@ -1,5 +1,0 @@
-import { Place } from '../models/Place'
-
-export interface PlaceRepository {
-  getPlaces(): Promise<Place[]>
-}
