+++ features/trips/domain/models/Place.ts
@@ -7,6 +7,11 @@ export type Place = {
     en_properties: string[]
     en_type: string
     id: string
+    lat: number
+    long: number
+    vi_name: string
+    vi_properties: string[]
+    vi_type: string
   }
   sort: string[]
 }
