+++ features/trips/infrastructure/api/PlaceApi.ts
@@ -1,24 +1,11 @@
-import { Place } from '@/domain/models/Place'
-import api, { ENDPOINTS } from './coreApi'
-
-interface GetPlacesParams {
-  limit?: number
-  location?: string
-  language?: string
-  filter?: string
-  search_after_id?: string
-}
+import { Place } from '@/features/trips/domain/models/Place'
+import { GetPlacesParams } from '../../domain/repositories/IPlaceRepository'
+import coreApi, { ENDPOINTS } from '../../../../lib/coreApi'
 
 export class PlaceApi {
-  static async getPlaces(
-    params: GetPlacesParams = {
-      limit: 1,
-      location: 'Ho Chi Minh',
-      language: 'en',
-    }
-  ): Promise<Place[]> {
+  static async getPlaces(params: GetPlacesParams): Promise<Place[]> {
     try {
-      const response = await api.get(ENDPOINTS.PLACES.BASE, {
+      const response = await coreApi.get(ENDPOINTS.PLACES.BASE, {
         params: {
           limit: params.limit,
           location: params.location,
@@ -27,7 +14,7 @@ export class PlaceApi {
           search_after_id: params.search_after_id,
         },
       })
-      return response.data
+      return response.data.data
     } catch (error) {
       console.error('Error fetching places:', error)
       throw error
@@ -36,7 +23,7 @@ export class PlaceApi {
 
   static async getPlaceById(id: string): Promise<Place> {
     try {
-      const response = await api.get(ENDPOINTS.PLACES.BY_ID(id))
+      const response = await coreApi.get(ENDPOINTS.PLACES.BY_ID(id))
       return response.data
     } catch (error) {
       console.error('Error fetching place by id:', error)
