+++ lib/beApi.ts
@@ -1,33 +1,21 @@
-import axios from 'axios'
 import { getItemAsync } from 'expo-secure-store'
+import createAxiosInstance from './axios'
 
-const api = axios.create({
-  timeout: 5000,
-  headers: { 'Content-Type': 'application/json' },
-})
+export const BE_URL =
+  process.env.EXPO_PUBLIC_BE_API_URL ?? 'http://localhost:3000/beApi/v1'
 
-// Log request before sending
-api.interceptors.request.use(
+const beApi = createAxiosInstance(BE_URL)
+
+// Add authentication interceptor
+beApi.interceptors.request.use(
   async (config) => {
-    // You can replace this with your own token retrieval logic
     const accessToken = await getItemAsync('accessToken')
-
     if (accessToken) {
       config.headers.Authorization = `Bearer ${accessToken}`
     }
-
-    console.log('Sending Request:', {
-      url: config.url,
-      method: config.method,
-      params: config.params,
-      data: config.data,
-      headers: config.headers,
-    })
-
     return config
   },
   (error) => {
-    console.error('Request Error:', error)
     return Promise.reject(
       error instanceof Error ? error : new Error(String(error))
     )
@@ -35,7 +23,7 @@ api.interceptors.request.use(
 )
 
 // Add response interceptor to handle 204 responses
-api.interceptors.response.use(
+beApi.interceptors.response.use(
   (response) => {
     console.log('Response received:', {
       status: response.status,
@@ -51,11 +39,8 @@ api.interceptors.response.use(
       return Promise.resolve({ status: 204, data: null })
     }
     console.error('Response Error:', error)
-    return Promise.reject(error)
+    return Promise.reject(new Error(error.message))
   }
 )
 
-export const url =
-  process.env.EXPO_PUBLIC_API_URL || 'http://localhost:3000/api/v1'
-
-export default api
+export default beApi
