+++ services/api/tools/ApiCurrency.ts
@@ -1,4 +1,4 @@
-import api from '../api'
+import beApi from '../../../lib/beApi'
 
 const url = process.env.EXPO_PUBLIC_CURRENCY_API_URL
 
@@ -7,7 +7,7 @@ export async function apiGetCurrencies(
   convertedCurrency: string
 ): Promise<number> {
   try {
-    const response = await api.get(`${url}/${originCurrency}.json`)
+    const response = await beApi.get(`${url}/${originCurrency}.json`)
     return response.data[originCurrency][convertedCurrency] || 0
   } catch (error) {
     console.error('Error fetching currency data:', error)
